new File for one method is too overkill
