--- src/GhConfig.cpp
@@ -1,18 +1,0 @@
-#include "GhConfig.h"
-
-pstring GetGhConfigDirectory()
-{
-	pstring directory = getenv("HOME");
-	directory << "/.config";
-
-	if (!DoesDirectoryExist(directory.c_str()))
-		mkdir(directory.c_str(), S_IRWXU | S_IRWXG | S_IROTH | S_IXOTH);
-
-	directory << "/AimTuxGH/";
-
-	if (!DoesDirectoryExist(directory.c_str()))
-		mkdir(directory.c_str(), S_IRWXU | S_IRWXG | S_IROTH | S_IXOTH);
-
-	return directory;
-}
-
