--- src/GhConfig.h
@@ -1,5 +1,0 @@
-#pragma once
-
-#include "config.h"
-
-pstring GetGhConfigDirectory();
