+++ src/config.cpp
@@ -82,3 +82,19 @@ std::vector<std::string> GetConfigs()
 
 	return configs;
 }
+
+pstring GetGhConfigDirectory()
+{
+	pstring directory = getenv("HOME");
+	directory << "/.config";
+
+	if (!DoesDirectoryExist(directory.c_str()))
+		mkdir(directory.c_str(), S_IRWXU | S_IRWXG | S_IROTH | S_IXOTH);
+
+	directory << "/AimTuxGH/";
+
+	if (!DoesDirectoryExist(directory.c_str()))
+		mkdir(directory.c_str(), S_IRWXU | S_IRWXG | S_IROTH | S_IXOTH);
+
+	return directory;
+}
