+++ src/config.h
@@ -24,3 +24,5 @@ std::vector<Config> GetConfigs(const char* directory);
 
 pstring GetConfigDirectory();
 std::vector<std::string> GetConfigs();
+
+pstring GetGhConfigDirectory();
